# Documentacion/CraftQuest_OpenAPI_MVP_v4.yaml
openapi: 3.0.3
info:
  title: CraftQuest API
  version: "4.0.0"
  description: >
    API MVP completo v4. Las respuestas se identifican por AnswerOptionId.
    A/B/C/D son etiquetas visuales generadas por intento y guardadas como snapshot.
servers:
  - url: https://api.craftquestai.com/api
    description: Producción
  - url: https://localhost:7080/api
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /quizzes:
    post:
      tags: [Quizzes]
      summary: Crear cuestionario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizRequest"
      responses:
        "201":
          description: Cuestionario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDto"
  /quizzes/{quizId}/questions:
    post:
      tags: [Quizzes]
      summary: Crear pregunta con respuestas por key e IDs generados por backend
      description: >
        El cliente envia answerOptions con clientKey y correctAnswerKeys.
        El backend crea AnswerOptionId reales y define las correctas en QuestionCorrectAnswerOptions.
      parameters:
        - $ref: "#/components/parameters/QuizId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Pregunta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDto"
  /quizzes/{quizId}/question-imports/process:
    post:
      tags: [Imports]
      summary: Procesar formato externo a CQIF v2
      description: >
        Recibe Excel, TXT, CSV, JSON, ZIP o rawText. No crea preguntas definitivas.
        Devuelve importId para preview, normalizacion con IA y confirmacion.
      parameters:
        - $ref: "#/components/parameters/QuizId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessImportRequest"
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                sourceType:
                  type: string
                  enum: [xlsx, txt, csv, json, zip]
                useAiNormalization:
                  type: boolean
      responses:
        "202":
          description: Importacion procesada o en proceso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionImportStatusDto"
  /question-imports/{importId}/preview:
    get:
      tags: [Imports]
      summary: Vista previa de importacion
      parameters:
        - $ref: "#/components/parameters/ImportId"
      responses:
        "200":
          description: Preview con preguntas, respuestas por key y errores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionImportPreviewDto"
  /question-imports/{importId}/ai-normalize:
    post:
      tags: [AI, Imports]
      summary: Normalizar con IA hacia CQIF v2
      parameters:
        - $ref: "#/components/parameters/ImportId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiNormalizeImportRequest"
      responses:
        "202":
          description: Job de IA creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiJobDto"
  /question-imports/{importId}/confirm:
    post:
      tags: [Imports]
      summary: Confirmar importacion y crear preguntas definitivas
      description: >
        Crea Questions, QuestionAnswerOptions y QuestionCorrectAnswerOptions.
        Las correctas se resuelven usando correctAnswerKeys de CQIF v2.
      parameters:
        - $ref: "#/components/parameters/ImportId"
      responses:
        "200":
          description: Importacion confirmada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionImportConfirmResultDto"
...
  /practice-sessions:
    post:
      tags: [Practice]
      summary: Iniciar sesion de practica
      description: >
        El backend materializa snapshots y aleatoriza respuestas. La respuesta no incluye isCorrect.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartPracticeSessionRequest"
      responses:
        "201":
          description: Sesion iniciada con preguntas y respuestas aleatorizadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSessionDto"
  /practice-sessions/{sessionId}/questions/{practiceQuestionSnapshotId}/answer:
    post:
      tags: [Practice]
      summary: Responder pregunta usando AnswerOptionId
      description: >
        El cliente debe enviar selectedAnswerOptionIds. No se aceptan letras visuales como A/B/C/D.
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/PracticeQuestionSnapshotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerRequest"
      responses:
        "200":
          description: Respuesta registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitAnswerResultDto"
  /practice-sessions/{sessionId}/finish:
    post:
      tags: [Practice]
      summary: Finalizar sesion de practica
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Resultado final
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSessionResultDto"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    QuizId:
      name: quizId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ImportId:
      name: importId
      in: path
      required: true
      schema: { type: string, format: uuid }
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PracticeQuestionSnapshotId:
      name: practiceQuestionSnapshotId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    CreateQuizRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 220 }
        description: { type: string }
        visibility:
          type: string
          enum: [private, shared_by_code, class_only, public, curated]
        randomizeQuestions: { type: boolean }
        defaultRandomizeAnswerOptions: { type: boolean }
    QuizDto:
      type: object
      properties:
        quizId: { type: string, format: uuid }
        title: { type: string }
        publicationStatus: { type: string }
    CreateQuestionRequest:
      type: object
      required: [questionType, text, answerOptions, correctAnswerKeys]
      properties:
        questionType:
          type: string
          enum: [single_choice, multiple_choice, true_false, image_choice, image_based_question]
        sectionId:
          type: string
          format: uuid
          nullable: true
        text:
          type: string
        points:
          type: number
          format: double
          default: 1
        randomizeAnswerOptions:
          type: boolean
          default: true
        scoringPolicy:
          type: string
          enum: [strict, partial_future]
          default: strict
        answerOptions:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/CreateAnswerOptionRequest"
        correctAnswerKeys:
          type: array
          minItems: 1
          items:
            type: string
        justification:
          $ref: "#/components/schemas/QuestionJustificationInput"
    CreateAnswerOptionRequest:
      type: object
      required: [clientKey]
      properties:
        clientKey:
          type: string
          description: Key estable en la solicitud/importacion; se convierte a StableKey.
        text:
          type: string
          nullable: true
        mediaAssetId:
          type: string
          format: uuid
          nullable: true
        defaultSortOrder:
          type: integer
    QuestionDto:
      type: object
      properties:
        questionId: { type: string, format: uuid }
        questionType: { type: string }
        text: { type: string }
        randomizeAnswerOptions: { type: boolean }
        answerOptions:
          type: array
          items:
            $ref: "#/components/schemas/AnswerOptionDto"
        correctAnswerOptionIds:
          type: array
          items: { type: string, format: uuid }
    AnswerOptionDto:
      type: object
      properties:
        answerOptionId: { type: string, format: uuid }
        stableKey: { type: string }
        text: { type: string }
        mediaAssetId: { type: string, format: uuid, nullable: true }
    QuestionJustificationInput:
      type: object
      properties:
        text: { type: string }
        sourceUrl: { type: string }
        visibility:
          type: string
          enum: [never, after_answer, after_quiz, teacher_only]
    ProcessImportRequest:
      type: object
      required: [sourceType]
      properties:
        sourceType:
          type: string
          enum: [xlsx, txt, csv, json, zip, raw_text]
        rawText:
          type: string
        useAiNormalization:
          type: boolean
        language:
          type: string
          default: es
    QuestionImportStatusDto:
      type: object
      properties:
        importId: { type: string, format: uuid }
        status: { type: string }
        totalQuestionsDetected: { type: integer }
        validQuestions: { type: integer }
        questionsWithWarnings: { type: integer }
        questionsWithErrors: { type: integer }
    QuestionImportPreviewDto:
      type: object
      properties:
        importId: { type: string, format: uuid }
        status: { type: string }
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CqifQuestion"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportErrorDto"
    ImportErrorDto:
      type: object
      properties:
        rowNumber: { type: integer }
        fieldName: { type: string }
        errorCode: { type: string }
        message: { type: string }
        severity:
          type: string
          enum: [info, warning, error]
    AiNormalizeImportRequest:
      type: object
      properties:
        generateMissingJustifications: { type: boolean, default: false }
        validateSemantics: { type: boolean, default: true }
        useGrounding: { type: boolean, default: false }
...
    AiJobDto:
      type: object
      properties:
        aiJobId: { type: string, format: uuid }
        status: { type: string }
        jobType: { type: string }
...
    CqifQuestion:
      type: object
      required: [type, text, answerOptions, correctAnswerKeys]
      properties:
        type:
          type: string
          enum: [single_choice, multiple_choice, true_false, image_choice, image_based_question]
        section: { type: string }
        text: { type: string }
        points: { type: number }
        randomizeAnswerOptions: { type: boolean }
        answerOptions:
          type: array
          items:
            $ref: "#/components/schemas/CqifAnswerOption"
        correctAnswerKeys:
          type: array
          items: { type: string }
        justification: { type: string }
    CqifAnswerOption:
      type: object
      required: [key]
      properties:
        key: { type: string }
        text: { type: string }
        mediaFileName: { type: string }
    QuestionImportConfirmResultDto:
      type: object
      properties:
        importId: { type: string, format: uuid }
        createdQuestions: { type: integer }
        skippedQuestions: { type: integer }
        createdQuestionIds:
          type: array
          items: { type: string, format: uuid }
    StartPracticeSessionRequest:
      type: object
      required: [quizId]
      properties:
        quizId: { type: string, format: uuid }
        classId: { type: string, format: uuid, nullable: true }
        assignmentId: { type: string, format: uuid, nullable: true }
        mode:
          type: string
          enum: [practice, exam, teacher_assignment]
          default: practice
    PracticeSessionDto:
      type: object
      properties:
        practiceSessionId: { type: string, format: uuid }
        quizId: { type: string, format: uuid }
        status: { type: string }
        questions:
          type: array
          items:
            $ref: "#/components/schemas/PracticeQuestionDto"
    PracticeQuestionDto:
      type: object
      properties:
        practiceQuestionSnapshotId: { type: string, format: uuid }
        questionId: { type: string, format: uuid }
        displayOrder: { type: integer }
        questionText: { type: string }
        questionType: { type: string }
        answers:
          type: array
          items:
            $ref: "#/components/schemas/PracticeAnswerOptionDto"
    PracticeAnswerOptionDto:
      type: object
      properties:
        answerOptionId: { type: string, format: uuid }
        displayOrder: { type: integer }
        displayLabel: { type: string, example: A }
        text: { type: string }
        mediaAssetId: { type: string, format: uuid, nullable: true }
    SubmitAnswerRequest:
      type: object
      required: [selectedAnswerOptionIds]
      properties:
        selectedAnswerOptionIds:
          type: array
          items: { type: string, format: uuid }
          description: IDs reales de respuestas seleccionadas. No enviar A/B/C/D.
    SubmitAnswerResultDto:
      type: object
      properties:
        practiceQuestionSnapshotId: { type: string, format: uuid }
        accepted: { type: boolean }
        answerStatus: { type: string }
    PracticeSessionResultDto:
      type: object
      properties:
        practiceSessionId: { type: string, format: uuid }
        scoreObtained: { type: number }
        scorePossible: { type: number }
        percentage: { type: number }
        correctAnswers: { type: integer }
        incorrectAnswers: { type: integer }
        omittedAnswers: { type: integer }
...